# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Drape API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /client/orders:
    get:
      operationId: getClientOrders
      tags: [client]
      summary: List all orders for the authenticated client
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientOrder"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /client/orders/{id}:
    get:
      operationId: getClientOrder
      tags: [client]
      summary: Get a single order with messages and review
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientOrderDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /client/orders/{id}/review:
    post:
      operationId: createOrderReview
      tags: [client]
      summary: Submit a review for a delivered order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderReview"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /client/messages/{orderId}:
    get:
      operationId: getOrderMessages
      tags: [client]
      summary: Get messages for an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Messages list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderMessage"
    post:
      operationId: sendOrderMessage
      tags: [client]
      summary: Send a message on an order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderMessage"
...
  /client/notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [client]
      summary: Mark a single notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
  /client/measurements:
    get:
      operationId: getClientMeasurements
      tags: [client]
      summary: Get saved body measurements
      responses:
        "200":
          description: Measurements or null
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ClientMeasurements"
                  - type: "null"
    put:
      operationId: updateClientMeasurements
      tags: [client]
      summary: Save or update body measurements
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMeasurementsInput"
      responses:
        "200":
          description: Saved measurements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientMeasurements"
  /client/profile:
    get:
      operationId: getClientStyleProfile
      tags: [client]
      summary: Get client style profile
      responses:
        "200":
          description: Style profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientStyleProfile"
    patch:
      operationId: updateClientStyleProfile
      tags: [client]
      summary: Update client style profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStyleProfileInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /producer/orders/{id}:
    get:
      operationId: getProducerOrder
      tags: [producer]
      summary: Get full order detail for producer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Full order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProducerOrderDetail"
    patch:
      operationId: updateProducerOrder
      tags: [producer]
      summary: Update order status, notes, or timeline
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProducerOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
    OrderStatus:
      type: string
      enum: [ENQUIRY, ACCEPTED, DEPOSIT_PAID, IN_PRODUCTION, FITTING, FINAL_PAYMENT, DELIVERED, COMPLETED, CANCELLED]
    TimelineEvent:
      type: object
      properties:
        date:
          type: string
        label:
          type: string
        completed:
          type: boolean
        note:
          type: ["string", "null"]
      required: [date, label, completed]
    ClientOrder:
      type: object
      properties:
        id:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        title:
          type: string
        description:
          type: ["string", "null"]
        agreedPrice:
          type: ["integer", "null"]
        currency:
          type: string
        depositPaid:
          type: boolean
        dueDate:
          type: ["string", "null"]
        estimatedDays:
          type: ["integer", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        producerId:
          type: string
        producerName:
          type: ["string", "null"]
        producerStudioName:
          type: ["string", "null"]
        reviewed:
          type: boolean
      required: [id, status, title, currency, depositPaid, createdAt, updatedAt, producerId, reviewed]
    OrderMessage:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        senderId:
          type: string
        senderRole:
          type: string
        content:
          type: string
        readByClient:
          type: boolean
        readByProducer:
          type: boolean
        createdAt:
          type: string
      required: [id, orderId, senderId, senderRole, content, readByClient, readByProducer, createdAt]
    OrderReview:
      type: object
      properties:
        id:
          type: string
        orderId:
          type: string
        clientId:
          type: string
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, orderId, clientId, rating, createdAt]
    ClientOrderDetail:
      allOf:
        - $ref: "#/components/schemas/ClientOrder"
        - type: object
          properties:
            notes:
              type: ["string", "null"]
            timelineEvents:
              type: array
              items:
                $ref: "#/components/schemas/TimelineEvent"
            briefId:
              type: ["string", "null"]
            sessionId:
              type: ["string", "null"]
            producerInstagram:
              type: ["string", "null"]
            producerWhatsapp:
              type: ["string", "null"]
            messages:
              type: array
              items:
                $ref: "#/components/schemas/OrderMessage"
            review:
              oneOf:
                - $ref: "#/components/schemas/OrderReview"
                - type: "null"
    CreateReviewInput:
      type: object
      properties:
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
      required: [rating]
    SendMessageInput:
      type: object
      properties:
        content:
          type: string
      required: [content]
...
    ClientMeasurements:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        unit:
          type: string
          enum: [cm, in]
        data:
          type: object
          additionalProperties:
            type: ["number", "null"]
        notes:
          type: ["string", "null"]
        updatedAt:
          type: string
        createdAt:
          type: string
      required: [id, userId, unit, data, updatedAt, createdAt]
    UpdateMeasurementsInput:
      type: object
      properties:
        unit:
          type: string
          enum: [cm, in]
        data:
          type: object
          additionalProperties:
            type: ["number", "null"]
        notes:
          type: string
      required: []
    ClientStyleProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        stylePreferences:
          type: array
          items:
            type: string
        budgetMin:
          type: ["integer", "null"]
        budgetMax:
          type: ["integer", "null"]
        styleNote:
          type: ["string", "null"]
        measurements:
          oneOf:
            - $ref: "#/components/schemas/ClientMeasurements"
            - type: "null"
      required: [id, email, stylePreferences]
    UpdateStyleProfileInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        city:
          type: string
        country:
          type: string
        bio:
          type: string
        stylePreferences:
          type: array
          items:
            type: string
        budgetMin:
          type: integer
        budgetMax:
          type: integer
        styleNote:
          type: string
      required: []
...
    ProducerOrderDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        status:
          type: string
        description:
          type: ["string", "null"]
        agreedPrice:
          type: ["integer", "null"]
        currency:
          type: string
        depositPaid:
          type: boolean
        dueDate:
          type: ["string", "null"]
        estimatedDays:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        timelineEvents:
          type: array
          items:
            type: object
        productionGuideAt:
          type: ["string", "null"]
        productionGuideContent:
          type: ["object", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
        client:
          type: ["object", "null"]
        measurements:
          type: ["object", "null"]
        brief:
          type: ["object", "null"]
        lookbookImages:
          type: array
          items:
            type: object
        messages:
          type: array
          items:
            type: object
      required: [id, title, status, currency, depositPaid, createdAt, updatedAt, lookbookImages, messages]
    UpdateProducerOrderBody:
      type: object
      properties:
        status:
          type: string
        notes:
          type: string
        timelineEvents:
          type: array
          items:
            type: object
        agreedPrice:
          type: integer
        depositPaid:
          type: boolean
        estimatedDays:
          type: integer
        dueDate:
          type: string
...